Crate antimony_monitor

This is a relatively simple implementation of a SECCOMP Notify Monitor. It doesn’t do anything more than log syscalls, permitting them all.

This utility is used while a profile is in Permissive mode. When in Enforcing, the logged syscalls are the only ones permitted, and the Filter is loaded immediately.

Structs

Cli

Enums

Error

Functions

audit_reader

Read from the Audit log

commit_or_defer

main 🔒

Receive and Respond to Notify Requests.

notify

notify_reader

update_binary 🔒

update_profile 🔒